fix bugs

+++ src/main.cpp
@@ -188,7 +188,7 @@ double get_simple_prediction(double s, double vx, double vy, int number_timestep
 	return predicted_s;
 }
 
-void get_prediction(double car_vx, double car_vy, double car_s, double car_d,
+void get_prediction(double car_vx, double car_vy, float car_s, float car_d,
 		const vector<double> &map_waypoints_s, const vector<double> &map_waypoints_x, const vector<double> &map_waypoints_y,
 		vector<double> &next_x_vals, vector<double> &next_y_vals) {
 
@@ -202,7 +202,7 @@ void get_prediction(double car_vx, double car_vy, double car_s, double car_d,
 	}
 }
 
-void get_trajectory(double car_x, double car_y, double car_yaw, double car_s, const vector<double> &previous_path_x, const vector<double> &previous_path_y,
+void get_trajectory(float car_x, float car_y, float car_yaw, float car_s, const vector<double> &previous_path_x, const vector<double> &previous_path_y,
 		const vector<double> &map_waypoints_s, const vector<double> &map_waypoints_x, const vector<double> &map_waypoints_y, const int lane, const double ref_vel,
 		vector<double> &next_x_vals, vector<double> &next_y_vals) {
